qkd-server/requirements.txt
hl7apy~=1.3.5
pydicom~=2.4.4
numpy~=1.26.4
fastapi~=0.111.0
pydantic~=2.7.4
pillow~=10.3.0
google-generativeai
langchain-community
streamlit
qiskit~=0.46.0
qiskit-aer~=0.14.0.1
pillow~=10.3.0
uvicorn
matplotlib
pylatexenc